ignore unused ctor props - support Java unit test templates as welli
