+++ src/integrationTest/java/com/weirddev/testme/intellij/generator/TestMeGeneratorGroovyJunit4Test.java
@@ -27,25 +27,11 @@ public class TestMeGeneratorGroovyJunit4Test extends TestMeGeneratorJunit4Test {
     public void testIgnoreUnusedProperties() throws Exception{
         doTest(true); //todo nested class getter call not identified - property someLongerNum
     }
-    public void testIgnoreUnusedCtorArguments() throws Exception{
-        doTest(true,true,true,67, true);
-    }
-    public void testIgnoreUnusedCtorArgumentsWhenDelegatedCalls() throws Exception{
-        doTest(true,true,true,67, true);
-    }
-//    public void testIgnoreUnusedCtorArgumentsWhenNestedProps() throws Exception{//todo support this use case
-//        doTest(true,true,true,67, true);
-//    }
-    public void testIgnoreUnusedCtorArgumentsInGroovy() throws Exception{
-        //note: 2nd ctor arg passed to BeanByCtor should actually be 'new Ice()' - rather than 'null' as currently set in excepted test outcome.
-        // For some reason manual tests match the expected behaviour but the UT fails. expected test has been adapted to the 'wrong' UT runtime behaviour
-        doTest(true,true,true,67, true);
+    public void testIgnoreUnusedPropertiesInGroovy() throws Exception {
+        doTest(true);
     }
     public void testDirectlyReferencedPropertiesNotIgnored() throws Exception{
         doTest("com.example.beans", "Foo", "FooTest", true, true, true, true, 50);
-    }
-    public void testIgnoreUnusedPropertiesInGroovy() throws Exception {
-        doTest(true);
     }
     public void testSettersOverCtor() throws Exception{
         doTest(true,true,true);
