+++ src/main/java/com/weirddev/testme/intellij/template/context/JavaTestBuilderImpl.java
@@ -3,6 +3,7 @@ package com.weirddev.testme.intellij.template.context;
 import com.intellij.openapi.diagnostic.Logger;
 import com.weirddev.testme.intellij.utils.ClassNameUtils;
 import com.weirddev.testme.intellij.utils.Node;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
@@ -20,17 +21,27 @@ public class JavaTestBuilderImpl implements LangTestBuilder {
     private static final Logger LOG = Logger.getInstance(JavaTestBuilderImpl.class.getName());
     private static final Pattern GENERICS_PATTERN = Pattern.compile("(<.*>)");
     private static Type DEFAULT_TYPE = new Type("java.lang.String", "String", "java.lang", false, false, false, false, false, new ArrayList<Type>());
+    public static final String PARAMS_SEPERATOR = ", ";
+
     protected final int maxRecursionDepth;
     protected final Method testedMethod;
+    /**
+     * do not invoke setter or getter for a property not used in tested method
+     */
+    protected final boolean shouldIgnoreUnusedProperties;
+    protected final TestBuilder.ParamRole paramRole; //todo consider removing. not used anymore
+    /**
+     * Relevant when shouldIgnoreUnusedProperties is true. When the minimum percentage of all interactions with constructed type are via setters/getters or direct property field read/assignment - then the type is considered as a 'data' bean,
+     * so a null value is passed for any constructor argument that is bound to a field in the constructed type which is not used in the tested method
+     */
+    protected final int minPercentOfInteractionWithPropertiesToTriggerConstructorOptimization;
 
-    public JavaTestBuilderImpl(int maxRecursionDepth) {
-        this.maxRecursionDepth = maxRecursionDepth;
-        testedMethod = null;
-    }
-
-    public JavaTestBuilderImpl(int maxRecursionDepth, Method testedMethod) {
+    public JavaTestBuilderImpl(int maxRecursionDepth, Method testedMethod, boolean shouldIgnoreUnusedProperties, TestBuilder.ParamRole paramRole, int minPercentOfInteractionWithPropertiesToTriggerConstructorOptimization) {
         this.maxRecursionDepth = maxRecursionDepth;
         this.testedMethod = testedMethod;
+        this.shouldIgnoreUnusedProperties = shouldIgnoreUnusedProperties;
+        this.paramRole = paramRole;
+        this.minPercentOfInteractionWithPropertiesToTriggerConstructorOptimization = minPercentOfInteractionWithPropertiesToTriggerConstructorOptimization;
     }
 
     //TODO consider aggregating conf into context object and managing maps outside of template
@@ -87,17 +98,6 @@ public class JavaTestBuilderImpl implements LangTestBuilder {
             testBuilder.append("}");
         }
     }
-
-    protected void buildCallParams(Method constructor, List<? extends Param> params, Map<String, String> replacementTypes, Map<String, String> defaultTypeValues, StringBuilder testBuilder, Node<Param> ownerParamNode) {
-        if (params != null) {
-            for (int i = 0; i < params.size(); i++) {
-                if (i != 0) {
-                    testBuilder.append(", ");
-                }
-                buildCallParam(replacementTypes, defaultTypeValues, testBuilder, new Node<Param>(params.get(i), ownerParamNode,ownerParamNode.getDepth()+1));
-            }
-        }
-    }
     protected void buildJavaParam(Map<String, String> replacementTypes, Map<String, String> defaultTypeValues, StringBuilder testBuilder, Node<Param> paramNode) {
         final Type type = paramNode.getData().getType();
         final String canonicalName = type.getCanonicalName();
@@ -150,6 +150,161 @@ public class JavaTestBuilderImpl implements LangTestBuilder {
             }
         }
     }
+    protected void buildCallParams(Method constructor, List<? extends Param> params, Map<String, String> replacementTypes, Map<String, String> defaultTypeValues, StringBuilder testBuilder, Node<Param> ownerParamNode) {
+        final int origLength = testBuilder.length();
+        if (params != null) {
+            final String ownerTypeCanonicalName = ownerParamNode.getData()==null?null:ownerParamNode.getData().getType().getCanonicalName();
+            boolean shouldOptimizeConstructorInitialization = ownerTypeCanonicalName !=null && constructor!=null && isShouldOptimizeConstructorInitialization(constructor,params, ownerTypeCanonicalName);
+            for (Param param : params) {
+                final Node<Param> paramNode = new Node<Param>(param, ownerParamNode, ownerParamNode.getDepth() + 1);
+                if (shouldIgnoreUnusedProperties && testedMethod != null) {
+                    if (isPropertyParam(paramNode.getData()) && ownerTypeCanonicalName != null && !isPropertyUsed(testedMethod, paramNode.getData(), ownerTypeCanonicalName)) {
+                        continue;
+                    } else if (shouldOptimizeConstructorInitialization && !param.getType().isPrimitive() && isUnused(testedMethod, deductAssignedToFields(constructor, param))) {
+                        testBuilder.append("null"+PARAMS_SEPERATOR);
+                        continue;
+                    }
+                }
+                buildCallParam(replacementTypes, defaultTypeValues, testBuilder, paramNode);
+                testBuilder.append(PARAMS_SEPERATOR);
+            }
+            if (origLength < testBuilder.length()) {
+                testBuilder.delete(testBuilder.length() - PARAMS_SEPERATOR.length(),testBuilder.length());
+            }
+        }
+    }
+    protected boolean isPropertyParam(Param param) {
+        return param instanceof SyntheticParam && ((SyntheticParam) param).isProperty;
+    }
+    protected boolean isShouldOptimizeConstructorInitialization(Method constructor, List<? extends Param> params, String ownerTypeCanonicalName) {
+        boolean shouldOptimizeConstructorInitialization = false;
+        if (shouldIgnoreUnusedProperties && testedMethod != null && params.size() > 0) {
+            int nBeanUsages = 0;
+            int nTotalTypeUsages = 0;
+            for (Param param : params) {
+                if (!isUnused(testedMethod, deductAssignedToFields(constructor,param))) {
+                    nBeanUsages++;
+                }
+            }
+            nTotalTypeUsages += countTypeReferences(ownerTypeCanonicalName, testedMethod);
+//            for (MethodCall methodCall : testedMethod.getCalledFamilyMembers()) {
+            for (MethodCall methodCall : testedMethod.getMethodCalls()) {
+                if (!methodCall.getMethod().isConstructor() && ownerTypeCanonicalName.equals(methodCall.getMethod().getOwnerClassCanonicalType())) {
+                    nTotalTypeUsages++;
+                }
+            }
+            shouldOptimizeConstructorInitialization = shouldOptimizeConstructorInitialization(nTotalTypeUsages, nBeanUsages);
+        }
+        return shouldOptimizeConstructorInitialization;
+    }
+    boolean shouldOptimizeConstructorInitialization(int nTotalTypeUsages, int nBeanUsages) {
+        return 0 < nBeanUsages && (nTotalTypeUsages * (minPercentOfInteractionWithPropertiesToTriggerConstructorOptimization / 100f) <= ((float)nBeanUsages) );
+    }
+
+    protected boolean isUnused(Method testedMethod, List<Field> fields) {
+        int unusedFieldsCount=0;
+        for (Field field : fields) {
+            if (!isPropertyUsed(testedMethod, new SyntheticParam(field.getType(),field.getName(),true),field.getOwnerClassCanonicalName())) {
+                unusedFieldsCount++;
+            }
+        }
+        if (fields.size() > 0 && fields.size() == unusedFieldsCount) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    protected boolean isPropertyUsed(@NotNull Method testedMethod, Param propertyParam, String paramOwnerCanonicalName) {
+        if (isReferencedInMethod(testedMethod, propertyParam, paramOwnerCanonicalName)) return true;
+        if (isPropertyUsedIndirectly(null, testedMethod, propertyParam, paramOwnerCanonicalName)) return true;
+//        for (MethodCall methodCall : testedMethod.getCalledFamilyMembers()) {
+        for (MethodCall methodCall : testedMethod.getMethodCalls()) {
+            if (isReferencedInMethod(methodCall.getMethod(), propertyParam, paramOwnerCanonicalName)) return true;
+            if (isPropertyUsedIndirectly(methodCall,methodCall.getMethod(), propertyParam, paramOwnerCanonicalName)) return true;
+        }
+        return false;
+    }
+    private boolean isPropertyUsedIndirectly(MethodCall methodCall, @NotNull Method method, Param propertyParam, String paramOwnerCanonicalName) {
+        if (methodCall !=null && paramOwnerCanonicalName.equals(methodCall.getMethod().getOwnerClassCanonicalType()) && isConstructorArgumentUsed(propertyParam, paramOwnerCanonicalName, methodCall, methodCall.getMethod())) {
+            return true;
+        }
+        for (MethodCall methodCallArg : method.getMethodCalls()) {
+            final Method methodCalled = methodCallArg.getMethod();
+            if (paramOwnerCanonicalName.equals(methodCalled.getOwnerClassCanonicalType()) && (isGetterUsed(propertyParam, methodCalled) || isSetterUsed(propertyParam, methodCalled) )) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private boolean isConstructorArgumentUsed(Param propertyParam, String paramOwnerCanonicalName, MethodCall methodCall, Method method) {
+        return method.isConstructor() && hasNonNullFieldMapping(methodCall, propertyParam,paramOwnerCanonicalName);
+    }
+
+    private boolean isSetterUsed(Param propertyParam, Method calledMethod) {
+        return calledMethod.isSetter() && calledMethod.getMethodParams().size()==1 && calledMethod.getMethodParams().get(0).getType().equals(propertyParam.getType()) && propertyParam.getName().equals(calledMethod.getPropertyName());
+    }
+
+    private boolean isGetterUsed(Param propertyParam, Method calledMethod) {
+        return calledMethod.isGetter() && calledMethod.getReturnType().getCanonicalName().equals(propertyParam.getType().getCanonicalName());
+    }
+    private boolean hasNonNullFieldMapping(MethodCall methodCall, Param propertyParam, String paramOwnerCanonicalName) {
+        for (int i = 0; i < methodCall.getMethod().getMethodParams().size(); i++) {
+            for (Field field : deductAssignedToFields(methodCall.getMethod(), methodCall.getMethod().getMethodParams().get(i))) {
+                if (methodCall.getMethodCallArguments() != null && methodCall.getMethodCallArguments().size() > i   && !"null".equals(methodCall.getMethodCallArguments().get(i).getText())
+                        && field.getName().equals(propertyParam.getName()) && field.getType().equals(propertyParam.getType()) && field.getOwnerClassCanonicalName().equals(paramOwnerCanonicalName)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+    private boolean isReferencedInMethod(@NotNull Method method, Param propertyParam, String paramOwnerTypeCanonicalName) {
+        if (method.isConstructor()) { /*assuming a more extensive ctor param usage check already done*/
+            return false;
+        }
+        for (Reference internalReference : method.getInternalReferences()) {
+            if (paramOwnerTypeCanonicalName.equals(internalReference.getOwnerType().getCanonicalName()) && propertyParam.getType().equals(internalReference.getReferenceType())
+                    && propertyParam.getName().equals(internalReference.getReferenceName())) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    protected List<Field> deductAssignedToFields(Method constructor, Param param) {
+        final ArrayList<Field> assignedToFields = param.getAssignedToFields();
+        if (assignedToFields.size() == 0 ) {
+            return deductAffectedFields(constructor,param);
+        } else {
+            return assignedToFields;
+        }
+    }
+
+    private List<Field> deductAffectedFields(Method constructor, Param param) {
+        final List<Field> affectedFields = new ArrayList<Field>();
+        for (Field field : constructor.getIndirectlyAffectedFields()) {
+            if (field.getType().equals(param.getType())) {
+                affectedFields.add(field);
+            }
+        }
+        return affectedFields;
+    }
+
+    private int countTypeReferences(String ownerTypeCanonicalName, Method method) {
+        int nTotalTypeReferences = 0;
+        if (method != null) {
+            for (Reference internalReference : method.getInternalReferences()) {
+                if (ownerTypeCanonicalName.equals(internalReference.getOwnerType().getCanonicalName())) {
+                    nTotalTypeReferences++;
+                }
+            }
+        }
+        return nTotalTypeReferences;
+    }
+
 
     protected String resolveNestedClassTypeName(String typeName) {
         return ClassNameUtils.extractClassName(typeName);
